+++ game_scripts/factories/psychlab/arbitrary_visuomotor_mapping_factory.lua
@@ -17,6 +17,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 
 local brady_konkle_oliva2008 = require 'datasets.brady_konkle_oliva2008'
 local game = require 'dmlab.system.game'
+local log = require 'common.log'
 local helpers = require 'common.helpers'
 local psychlab_factory = require 'factories.psychlab.factory'
 local psychlab_helpers = require 'factories.psychlab.helpers'
@@ -230,10 +231,9 @@ function factory.createLevelApi(kwargs)
 
   -- 'init' gets called at the start of each episode.
   function env:_init(pac, opts)
-    print('opts passed to _init:')
-    print(helpers.tostring(opts))
-    print('ARG in _init:')
-    print(helpers.tostring(ARG))
+    log.info('opts passed to _init:\n' .. helpers.tostring(opts))
+    log.info('ARG passed to _init:\n' .. helpers.tostring(ARG))
+
 
     if self.dataset == nil then
       self.dataset = brady_konkle_oliva2008(opts)
@@ -244,7 +244,7 @@ function factory.createLevelApi(kwargs)
     -- If requested, randomly perturb the target location for each trial.
     self.jitter = ARG.jitter
     if self.jitter then
-      print('Jitter target location')
+      log.info('Jitter target location')
       self.jitteredCenter = {}
     end