+++ game_scripts/levels/contributed/psychlab/factories/visual_search_factory.lua
@@ -16,6 +16,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 ]]
 
 local game = require 'dmlab.system.game'
+local log = require 'common.log'
 local psychlab_factory = require 'factories.psychlab.factory'
 local psychlab_helpers = require 'factories.psychlab.helpers'
 local helpers = require 'common.helpers'
@@ -192,16 +193,15 @@ function factory.createLevelApi(kwargs)
 
   -- 'init' gets called at the start of each episode.
   function env:_init(pac, opts)
-    print('opts passed to _init:')
-    print(helpers.tostring(opts))
+    log.info('opts passed to _init:\n' .. helpers.tostring(opts))
+
     ARG.searchMode = opts.searchMode or 'interleaved'
     ARG.jitter = opts.jitter or false
 
-    print('ARGs in _init:')
-    print(helpers.tostring(ARG))
+    log.info('ARGs in _init:\n' .. helpers.tostring(ARG))
 
     ARG = parseArgs(ARG)
-    print('visual_search args parsed correctly')
+    log.info('visual_search args parsed correctly')
 
     self.screenSize = opts.screenSize
 
@@ -214,7 +214,7 @@ function factory.createLevelApi(kwargs)
     -- If requested, randomly perturb the target location for each trial.
     self.jitter = ARG.jitter
     if self.jitter then
-      print('Jitter target location')
+      log.info('Jitter target location')
       self.jitteredCenter = {}
     end